edited order model, created custom admin page and permissions

+++ cart_app/views.py
@@ -2,11 +2,27 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 from django.urls import reverse
 
-from cart_app.forms import OrderForm
+from cart_app.forms import OrderForm, EditItem
 from cart_app.models import Cart, CartItem, Order
 
 from product_app.models import ProductModel
 from the_hive_core.decorators import group_required
+
+
+def perm_check(request, group_names, current_user):
+    user_group = request.user.groups.values_list('name', flat=True)
+
+    if set(group_names).intersection(user_group):
+        return True
+
+    elif request.user.is_superuser:
+        return True
+
+    elif request.user == current_user:
+        return True
+
+    else:
+        return False
 
 
 @login_required(login_url='login')
@@ -16,7 +32,7 @@ def cart_view(request):
     cart_items = cart.items.all()
 
     context = {
-        'cart_items': cart_items,
+        'cart_items': cart_items.order_by('pk'),
         'total_price': cart.total_price()
     }
 
@@ -93,7 +109,6 @@ def checkout_page(request):
 
     cart = Cart.objects.get(user=request.user)
     cart_items = cart.items.all()
-    total_qty = sum(item.quantity for item in cart_items)
     delivery = 5
 
     form = OrderForm()
@@ -107,8 +122,6 @@ def checkout_page(request):
             order = form.save(commit=False)
 
             order.user = request.user
-            order.total_price = cart.total_price() + delivery
-            order.total_products_qty = total_qty
             order.save()
 
             for item in cart_items:
@@ -145,6 +158,10 @@ def successful_order_page(request):
 def delete_order(request, pk):
 
     order = Order.objects.get(pk=pk)
+
+    if not perm_check(request, ['Admin'], order.user):
+        return redirect('index')
+
     order.delete()
     url = reverse('profile-orders', kwargs={'pk': request.user.pk})
 
@@ -155,11 +172,57 @@ def delete_order(request, pk):
 def order_details(request, pk):
 
     order = Order.objects.get(pk=pk)
+
+    if not perm_check(request, ['Admin', 'Moderator'], order.user):
+        return redirect('index')
+
     order_items = order.items.all()
 
     context = {
         'order_items': order_items,
-        'order': order
+        'order': order,
+        'is_admin': request.user.groups.filter(name='Admin').exists(),
+        'is_moderator': request.user.groups.filter(name='Moderator').exists()
     }
 
     return render(request, 'cart/order_details.html', context)
+
+
+@login_required(login_url='login')
+def edit_item(request, pk):
+
+    item = CartItem.objects.get(pk=pk)
+    order = item.order_set.first()
+    form = EditItem(instance=item)
+
+    if not perm_check(request, ['Admin', 'Moderator'], None):
+        return redirect('index')
+
+    if request.method == 'POST':
+        form = EditItem(request.POST, instance=item)
+
+        if form.is_valid():
+            form.save()
+
+            url = reverse('order-details', kwargs={'pk': order.pk})
+            return redirect(url)
+
+    context = {
+        'form': form,
+        'item': item
+    }
+    return render(request, 'cart/item_edit.html', context)
+
+
+@login_required(login_url='login')
+def delete_item(request, pk):
+
+    item = CartItem.objects.get(pk=pk)
+
+    if not perm_check(request, ['Admin'], None):
+        return redirect('index')
+
+    item.delete()
+
+    previous_page = request.META.get('HTTP_REFERER')
+    return redirect(previous_page)
